GuidedTours: fix element highlighting
Element highlighting fades out Calypso except for the target element
the current step is about. This helps the user identify and focus on
the element the step is talking about.

This is currently used in `GuidesActionStep` steps, but should also be
used in all `GuidesBasicStep` steps that have a target.

The current approach doesn't seem to work in all cases where we need
it. Specifically it only fades out the sidebar if the target is an
element within the sidebar.

This commit adds a `GuidesOverlay` component that provides a more
robust overlay and is now used by both `GuidesBasicStep` and
`GuidesActionStep`.